Made the playing now icon animated
